Adding logging to WebMention::Client
Prior to this commit, when I called `Webmention::Client#send_all_mentions`,
I had no idea which URL was throwing an exception.

With logging, I was able to track down that the `Webmention::Client`
considered a `mailto` URL as something to attempt to send a mention
to.

By adding the logging, I can see the candidate URLs for sending a
webmention. I can also see if the candidate URL has a webmention
end point or not.